hireportai/.claude/worktrees/xenodochial-pike/hirelens-backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Document parsing
pdfplumber>=0.10.3
python-docx>=1.1.0

# NLP
spacy>=3.7.4
scikit-learn>=1.4.0
numpy>=1.26.3

# Database
sqlalchemy[asyncio]>=2.0.36
alembic>=1.13.1
aiosqlite>=0.19.0
aiofiles>=23.2.1

# Authentication
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Payments
stripe>=8.0.0

# LLM Providers
google-genai>=1.0.0
anthropic>=0.39.0
